docs: add TSTyche to the References section (#590)
## Overview

I would lie to propose adding TSTyche (a project of mine) to the
References list.

---

TSTyche is relatively new type testing tool. It felt like this project
is trying to mention all actively maintained alternative libraries. So
here is one more (;

I have created a separate section, because TSTyche is not just a “type
assertion library”. It is a type test runner. In my opinion, only tsd is
providing similar feature. But is far far far ... behind. For instance,
tsd was designed to test hand crafted `.d.ts` declarations, but unit
type testing or mono repo support is not implemented and not event
planned. Oh well.. (;

---

Feel free to adjust the text. I just wanted to updated the References
list (;

---------

Co-authored-by: Josh Goldberg ✨ <git@joshuakgoldberg.com>